requirements.txt
Django==4.0
psycopg2-binary==2.9.2
Pillow==8.4.0
django-constance[database]
